+++ jqm-all/jqm-api/src/main/java/com/enioka/jqm/api/JobBase.java
@@ -61,21 +61,21 @@ public class JobBase
         jm.sendProgress(progress);
     }
 
-    public int enQueue(String applicationName, String user, String mail, String sessionID, String application, String module,
-            String keyword1, String keyword2, String keyword3, Map<String, String> parameters)
+    public Long enQueue(String applicationName, String user, String mail, String sessionID, String application, String module,
+                        String keyword1, String keyword2, String keyword3, Map<String, String> parameters)
     {
         return jm.enqueue(applicationName, user, mail, sessionID, application, module, keyword1, keyword2, keyword3, parameters);
     }
 
-    public int enQueueSynchronously(String applicationName, String user, String mail, String sessionID, String application, String module,
-            String keyword1, String keyword2, String keyword3, Map<String, String> parameters)
+    public Long enQueueSynchronously(String applicationName, String user, String mail, String sessionID, String application, String module,
+                                     String keyword1, String keyword2, String keyword3, Map<String, String> parameters)
     {
         return jm.enqueueSync(applicationName, user, mail, sessionID, application, module, keyword1, keyword2, keyword3, parameters);
     }
 
     // ---------
 
-    public Integer getParentID()
+    public Long getParentID()
     {
         return jm.parentID();
     }
@@ -145,7 +145,7 @@ public class JobBase
         return jm.defaultConnect();
     }
 
-    public Integer getJobInstanceId()
+    public Long getJobInstanceId()
     {
         return jm.jobInstanceID();
     }
