+++ jqm-all/jqm-tst-osgi/src/main/java/com/enioka/jqm/tester/JqmSynchronousTesterOsgi.java
@@ -42,9 +42,9 @@ public class JqmSynchronousTesterOsgi implements JqmSynchronousTester
     private JqmSingleRunnerOperations runner;
 
     private Node node = null;
-    private Integer jd = null;
-    private Integer q = null;
-    private Integer ji = null;
+    private Long jd = null;
+    private Long q = null;
+    private Long ji = null;
 
     // Just for dependency.
     @Reference
