+++ jqm-all/jqm-ws/src/main/java/com/enioka/jqm/ws/api/ServiceClient.java
@@ -104,7 +104,7 @@ public class ServiceClient
             target.startHeld();
         }
 
-        int i = target.enqueue();
+        long i = target.enqueue();
 
         JobInstance ji = new JobInstance();
         ji.setId(i);
@@ -131,68 +131,68 @@ public class ServiceClient
 
     @Path("ji/{id}")
     @POST
-    public int enqueueFromHistory(@PathParam("id") int jobIdToCopy)
+    public Long enqueueFromHistory(@PathParam("id") long jobIdToCopy)
     {
         return JqmClientFactory.getClient().enqueueFromHistory(jobIdToCopy);
     }
 
     @Path("ji/cancelled/{jobId}")
     @POST
-    public void cancelJob(@PathParam("jobId") int jobId)
+    public void cancelJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().cancelJob(jobId);
     }
 
     @Path("ji/waiting/{jobId}")
     @DELETE
-    public void deleteJob(@PathParam("jobId") int jobId)
+    public void deleteJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().deleteJob(jobId);
     }
 
     @Path("ji/killed/{jobId}")
     @POST
-    public void killJob(@PathParam("jobId") int jobId)
+    public void killJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().killJob(jobId);
     }
 
     @Path("schedule/{scheduleId}")
     @DELETE
-    public void removeRecurrence(@PathParam("scheduleId") int scheduleId)
+    public void removeRecurrence(@PathParam("scheduleId") long scheduleId)
     {
         JqmClientFactory.getClient().removeRecurrence(scheduleId);
     }
 
     @Path("ji/paused/{jobId}")
     @POST
-    public void pauseQueuedJob(@PathParam("jobId") int jobId)
+    public void pauseQueuedJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().pauseQueuedJob(jobId);
     }
 
     @Path("ji/paused/{jobId}")
     @DELETE
-    public void resumeQueuedJob(@PathParam("jobId") int jobId)
+    public void resumeQueuedJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().resumeQueuedJob(jobId);
     }
 
-    public void resumeJob(@PathParam("jobId") int jobId)
+    public void resumeJob(@PathParam("jobId") long jobId)
     {
         resumeQueuedJob(jobId);
     }
 
     @Path("ji/running/paused/{jobId}")
     @POST
-    public void pauseRunningJob(@PathParam("jobId") int jobId)
+    public void pauseRunningJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().pauseRunningJob(jobId);
     }
 
     @Path("ji/running/paused/{jobId}")
     @DELETE
-    public void resumeRunningJob(@PathParam("jobId") int jobId)
+    public void resumeRunningJob(@PathParam("jobId") long jobId)
     {
         JqmClientFactory.getClient().resumeRunningJob(jobId);
     }
@@ -206,48 +206,48 @@ public class ServiceClient
     @Path("ji/crashed/{jobId}")
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @DELETE
-    public JobInstance restartCrashedJobObject(@PathParam("jobId") int jobId)
+    public JobInstance restartCrashedJobObject(@PathParam("jobId") long jobId)
     {
-        int i = JqmClientFactory.getClient().restartCrashedJob(jobId);
+        long i = JqmClientFactory.getClient().restartCrashedJob(jobId);
         return getJob(i);
     }
 
     @Path("q/{queueId: [0-9]+}/{jobId: [0-9]+}")
     @POST
-    public void setJobQueue(@PathParam("jobId") int jobId, @PathParam("queueId") int queueId)
+    public void setJobQueue(@PathParam("jobId") long jobId, @PathParam("queueId") long queueId)
     {
         JqmClientFactory.getClient().setJobQueue(jobId, queueId);
     }
 
     // No need to expose. Client side work.
 
-    public void setJobQueue(int jobId, Queue queue)
+    public void setJobQueue(Long jobId, Queue queue)
     {
         JqmClientFactory.getClient().setJobQueue(jobId, queue);
     }
 
     @POST
     @Path("ji/{jobId}/position/{newPosition}")
-    public void setJobQueuePosition(@PathParam("jobId") int jobId, @PathParam("newPosition") int newPosition)
+    public void setJobQueuePosition(@PathParam("jobId") long jobId, @PathParam("newPosition") int newPosition)
     {
         JqmClientFactory.getClient().setJobQueuePosition(jobId, newPosition);
     }
 
     @POST
     @Path("ji/{jobId}/priority/{priority}")
-    public void setJobPriority(@PathParam("jobId") int jobId, @PathParam("priority") int priority)
+    public void setJobPriority(@PathParam("jobId") long jobId, @PathParam("priority") int priority)
     {
         JqmClientFactory.getClient().setJobPriority(jobId, priority);
     }
 
-    public void setJobRunAfter(@PathParam("jobId") int jobId, @PathParam("whenToRun") Calendar whenToRun)
+    public void setJobRunAfter(@PathParam("jobId") long jobId, @PathParam("whenToRun") Calendar whenToRun)
     {
         JqmClientFactory.getClient().setJobRunAfter(jobId, whenToRun);
     }
 
     @POST
     @Path("ji/{jobId}/delay/{whenToRun}")
-    public void setJobRunAfter(@PathParam("jobId") int jobId, @PathParam("whenToRun") long whenToRun)
+    public void setJobRunAfter(@PathParam("jobId") long jobId, @PathParam("whenToRun") long whenToRun)
     {
         Calendar c = Calendar.getInstance();
         c.setTimeInMillis(whenToRun);
@@ -256,21 +256,21 @@ public class ServiceClient
 
     @POST
     @Path("schedule/{scheduleId}/queue/{queueId}")
-    public void setScheduleQueue(@PathParam("scheduleId") int scheduleId, @PathParam("queueId") int queueId)
+    public void setScheduleQueue(@PathParam("scheduleId") long scheduleId, @PathParam("queueId") long queueId)
     {
         JqmClientFactory.getClient().setScheduleQueue(scheduleId, queueId);
     }
 
     @POST
     @Path("schedule/{scheduleId}/cron/{cronExpression}")
-    public void setScheduleRecurrence(@PathParam("scheduleId") int scheduleId, @PathParam("cronExpression") String cronExpression)
+    public void setScheduleRecurrence(@PathParam("scheduleId") long scheduleId, @PathParam("cronExpression") String cronExpression)
     {
         JqmClientFactory.getClient().setScheduleRecurrence(scheduleId, cronExpression);
     }
 
     @POST
     @Path("schedule/{scheduleId}/priority/{priority}")
-    public void setSchedulePriority(@PathParam("scheduleId") int scheduleId, @PathParam("priority") int priority)
+    public void setSchedulePriority(@PathParam("scheduleId") long scheduleId, @PathParam("priority") int priority)
     {
         JqmClientFactory.getClient().setSchedulePriority(scheduleId, priority);
     }
@@ -279,7 +279,7 @@ public class ServiceClient
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @Path("ji/{jobId}")
     @HttpCache("public, max-age=60")
-    public JobInstance getJob(@PathParam("jobId") int jobId)
+    public JobInstance getJob(@PathParam("jobId") long jobId)
     {
         return JqmClientFactory.getClient().getJob(jobId);
     }
@@ -358,14 +358,14 @@ public class ServiceClient
     @GET
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @HttpCache("public, max-age=60")
-    public List<String> getJobMessages(@PathParam("jobId") int jobId)
+    public List<String> getJobMessages(@PathParam("jobId") long jobId)
     {
         return JqmClientFactory.getClient().getJobMessages(jobId);
     }
 
     // Not exposed. Use getJob => progress
 
-    public int getJobProgress(int jobId)
+    public int getJobProgress(long jobId)
     {
         throw new NotSupportedException();
     }
@@ -374,7 +374,7 @@ public class ServiceClient
     @GET
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @HttpCache("public, max-age=60")
-    public List<Deliverable> getJobDeliverables(@PathParam("jobId") int jobId)
+    public List<Deliverable> getJobDeliverables(@PathParam("jobId") long jobId)
     {
         return JqmClientFactory.getClient().getJobDeliverables(jobId);
     }
@@ -382,7 +382,7 @@ public class ServiceClient
     // Not exposed. Returning a list of files is a joke anyway... Loop should be
     // client-side.
 
-    public List<InputStream> getJobDeliverablesContent(int jobId)
+    public List<InputStream> getJobDeliverablesContent(long jobId)
     {
         throw new NotSupportedException();
     }
@@ -401,7 +401,7 @@ public class ServiceClient
     @Path("ji/files/{id}")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getDeliverableContent(@PathParam("id") int delId, @Context HttpServletResponse res)
+    public InputStream getDeliverableContent(@PathParam("id") long delId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) JqmClientFactory.getClient().getDeliverableContent(delId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -411,7 +411,7 @@ public class ServiceClient
     @Path("ji/{jobId}/stderr")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getJobLogStdErr(@PathParam("jobId") int jobId, @Context HttpServletResponse res)
+    public InputStream getJobLogStdErr(@PathParam("jobId") long jobId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) JqmClientFactory.getClient().getJobLogStdErr(jobId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -421,7 +421,7 @@ public class ServiceClient
     @Path("ji/{jobId}/stdout")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getJobLogStdOut(@PathParam("jobId") int jobId, @Context HttpServletResponse res)
+    public InputStream getJobLogStdOut(@PathParam("jobId") long jobId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) JqmClientFactory.getClient().getJobLogStdOut(jobId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -444,7 +444,7 @@ public class ServiceClient
 
     @Path("q/{qId}/pause")
     @POST
-    public void pauseQueue(@PathParam("qId") int qId)
+    public void pauseQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -458,7 +458,7 @@ public class ServiceClient
 
     @Path("q/{qId}/pause")
     @DELETE
-    public void resumeQueue(@PathParam("qId") int qId)
+    public void resumeQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -472,7 +472,7 @@ public class ServiceClient
 
     @Path("q/{qId}/clear")
     @POST
-    public void clearQueue(@PathParam("qId") int qId)
+    public void clearQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -481,7 +481,7 @@ public class ServiceClient
 
     @Path("q/{qId}/status")
     @GET
-    public QueueStatus getQueueStatus(@PathParam("qId") int qId)
+    public QueueStatus getQueueStatus(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -495,7 +495,7 @@ public class ServiceClient
 
     @Path("q/{qId}/enabled-capacity")
     @GET
-    public int getQueueEnabledCapacity(@PathParam("qId") int qId)
+    public int getQueueEnabledCapacity(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
